requirements.txt
pillow==10.3.0
Flask == 3.0.2
generativeai==                 0.0.1
google-ai-generativelanguage== 0.4.0
google-api-core  ==            2.18.0
google-auth ==                 2.29.0
google-cloud-vision ==          3.7.2
google-generativeai ==          0.4.1
google-pasta    ==             0.2.0
googleapis-common-protos ==    1.63.0